% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_15plus.R
\name{add_BMI}
\alias{add_BMI}
\title{add BMI and BMIclass to datatable}
\usage{
add_BMI(df, dM, reference_date)
}
\arguments{
\item{df}{dataframe
needs DOB, Sex
  HeightDate, HeightValue, WeightValue}

\item{dM}{dMeasure object}

\item{reference_date}{date to calculate age}
}
\value{
dataframe with added
    BMIValue, BMIDate, BMIClass
}
\description{
calculation of BMIClass is Age and Sex specific
}
\details{
see PIP-QI-Technical-Specifications v1.1 (Australian Government)
https://www1.health.gov.au/internet/main/publishing.nsf/
  Content/46506AF50A4824B6CA25848600113FFF/$File/PIP-QI-Technical-Specifications.pdf
}
