% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityImprovementMeasures.R
\name{add_demographics}
\alias{add_demographics}
\title{add demographics}
\usage{
add_demographics(df, dM, reference_date)
}
\arguments{
\item{df}{dataframe with InternalID}

\item{dM}{dMeasure object}

\item{reference_date}{date to calculate age from}
}
\value{
dataframe with DOB, Age10, Sex, Indigenous, RecordNo,
  MaritalStatus, Sexuality, Indigenous
}
\description{
add demographics
}
