% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{datatableUI}
\alias{datatableUI}
\title{QIM module - UI function}
\usage{
datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
Display appointments within selected range of dates and providers
}
