% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityImprovementMeasures.R
\name{dMeasureIntegration}
\alias{dMeasureIntegration}
\title{dMeasureIntegration}
\usage{
dMeasureIntegration(information)
}
\arguments{
\item{information}{the information required
`Provides` - modules provided (in this case, `dMeasureQIMRept` and `dMeasureQIMAppt`)
`Requires` - the modules required (including `dMeasure`)
`moduleID` - IDs of modules to create}
}
\value{
vector of required information
}
\description{
integration with dMeasure
  (especially DailyMeasure)
}
