% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityImprovementMeasures.R
\name{dMeasureQIM}
\alias{dMeasureQIM}
\title{dMeasureQIM class}
\description{
Quality Improvement Measures reports and case-finding in EMR (Best Practice)
    a module for dMeasure
}
\details{
dMeasureQIM class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-list_qim_15plus}{\code{dMeasureQIM$list_qim_15plus()}}
\item \href{#method-list_qim_15plus_appointments}{\code{dMeasureQIM$list_qim_15plus_appointments()}}
\item \href{#method-report_qim_15plus}{\code{dMeasureQIM$report_qim_15plus()}}
\item \href{#method-list_qim_65plus}{\code{dMeasureQIM$list_qim_65plus()}}
\item \href{#method-list_qim_65plus_appointments}{\code{dMeasureQIM$list_qim_65plus_appointments()}}
\item \href{#method-report_qim_65plus}{\code{dMeasureQIM$report_qim_65plus()}}
\item \href{#method-list_qim_active}{\code{dMeasureQIM$list_qim_active()}}
\item \href{#method-report_qim_active}{\code{dMeasureQIM$report_qim_active()}}
\item \href{#method-list_qim_copd}{\code{dMeasureQIM$list_qim_copd()}}
\item \href{#method-list_qim_copd_appointments}{\code{dMeasureQIM$list_qim_copd_appointments()}}
\item \href{#method-report_qim_copd}{\code{dMeasureQIM$report_qim_copd()}}
\item \href{#method-list_qim_cst}{\code{dMeasureQIM$list_qim_cst()}}
\item \href{#method-list_qim_cst_appointments}{\code{dMeasureQIM$list_qim_cst_appointments()}}
\item \href{#method-report_qim_cst}{\code{dMeasureQIM$report_qim_cst()}}
\item \href{#method-list_qim_cvdRisk}{\code{dMeasureQIM$list_qim_cvdRisk()}}
\item \href{#method-list_qim_cvdRisk_appointments}{\code{dMeasureQIM$list_qim_cvdRisk_appointments()}}
\item \href{#method-report_qim_cvdRisk}{\code{dMeasureQIM$report_qim_cvdRisk()}}
\item \href{#method-list_qim_diabetes}{\code{dMeasureQIM$list_qim_diabetes()}}
\item \href{#method-list_qim_diabetes_appointments}{\code{dMeasureQIM$list_qim_diabetes_appointments()}}
\item \href{#method-report_qim_diabetes}{\code{dMeasureQIM$report_qim_diabetes()}}
\item \href{#method-new}{\code{dMeasureQIM$new()}}
\item \href{#method-clone}{\code{dMeasureQIM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_15plus"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_15plus}{}}}
\subsection{Method \code{list_qim_15plus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_15plus(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_15plus_appointments"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_15plus_appointments}{}}}
\subsection{Method \code{list_qim_15plus_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_15plus_appointments(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report_qim_15plus"></a>}}
\if{latex}{\out{\hypertarget{method-report_qim_15plus}{}}}
\subsection{Method \code{report_qim_15plus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$report_qim_15plus(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  demographic = NA,
  measure = NA,
  ignoreOld = NA,
  lazy = FALSE,
  store = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_65plus"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_65plus}{}}}
\subsection{Method \code{list_qim_65plus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_65plus(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_65plus_appointments"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_65plus_appointments}{}}}
\subsection{Method \code{list_qim_65plus_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_65plus_appointments(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report_qim_65plus"></a>}}
\if{latex}{\out{\hypertarget{method-report_qim_65plus}{}}}
\subsection{Method \code{report_qim_65plus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$report_qim_65plus(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  demographic = NA,
  ignoreOld = NA,
  lazy = FALSE,
  store = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_active"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_active}{}}}
\subsection{Method \code{list_qim_active()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_active(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report_qim_active"></a>}}
\if{latex}{\out{\hypertarget{method-report_qim_active}{}}}
\subsection{Method \code{report_qim_active()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$report_qim_active(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  demographic = NA,
  lazy = FALSE,
  store = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_copd"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_copd}{}}}
\subsection{Method \code{list_qim_copd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_copd(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_copd_appointments"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_copd_appointments}{}}}
\subsection{Method \code{list_qim_copd_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_copd_appointments(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report_qim_copd"></a>}}
\if{latex}{\out{\hypertarget{method-report_qim_copd}{}}}
\subsection{Method \code{report_qim_copd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$report_qim_copd(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  demographic = NA,
  ignoreOld = NA,
  lazy = FALSE,
  store = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_cst"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_cst}{}}}
\subsection{Method \code{list_qim_cst()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_cst(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_cst_appointments"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_cst_appointments}{}}}
\subsection{Method \code{list_qim_cst_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_cst_appointments(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report_qim_cst"></a>}}
\if{latex}{\out{\hypertarget{method-report_qim_cst}{}}}
\subsection{Method \code{report_qim_cst()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$report_qim_cst(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  demographic = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_cvdRisk"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_cvdRisk}{}}}
\subsection{Method \code{list_qim_cvdRisk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_cvdRisk(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_cvdRisk_appointments"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_cvdRisk_appointments}{}}}
\subsection{Method \code{list_qim_cvdRisk_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_cvdRisk_appointments(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report_qim_cvdRisk"></a>}}
\if{latex}{\out{\hypertarget{method-report_qim_cvdRisk}{}}}
\subsection{Method \code{report_qim_cvdRisk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$report_qim_cvdRisk(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  demographic = NA,
  ignoreOld = NA,
  lazy = FALSE,
  store = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_diabetes"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_diabetes}{}}}
\subsection{Method \code{list_qim_diabetes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_diabetes(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_qim_diabetes_appointments"></a>}}
\if{latex}{\out{\hypertarget{method-list_qim_diabetes_appointments}{}}}
\subsection{Method \code{list_qim_diabetes_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$list_qim_diabetes_appointments(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report_qim_diabetes"></a>}}
\if{latex}{\out{\hypertarget{method-report_qim_diabetes}{}}}
\subsection{Method \code{report_qim_diabetes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$report_qim_diabetes(
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  demographic = NA,
  measure = NA,
  ignoreOld = NA,
  lazy = FALSE,
  store = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$new(dMeasure_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureQIM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
