% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityImprovementMeasures.R
\name{fill_demographics}
\alias{fill_demographics}
\title{fill report with 'complete' quality improvement measures and demographics}
\usage{
fill_demographics(d)
}
\arguments{
\item{d}{dataframe to process
with QIM, Age10, Sex, Indigenous, DiabetesType
Measure, State, n.
optionally - ProportionDemographic, DateFrom, DateTo}
}
\value{
dataframe with DOB, Age10, Sex, Indigenous, DiabetesType
  Measure, State, n.
  optionally - ProportionDemographic, DateFrom, DateTo
}
\description{
determines from the content of 'QIM' what is required to fill
in the 'missing' (zero) rows
}
