% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_report.R
\name{getReport}
\alias{getReport}
\title{getReport}
\usage{
getReport(
  report_function,
  date_from,
  date_to,
  contact_type,
  min_contact,
  progress = NA,
  progress_detail,
  measure = NA,
  require_type_diabetes,
  qim_name,
  measure_name,
  state_variable_name,
  small_cell_suppression = FALSE,
  include_all_demographic_groups = FALSE,
  age_min = 0,
  age_max = 65,
  states = c(FALSE, TRUE)
)
}
\arguments{
\item{report_function}{the dMQIM method to call}

\item{date_from}{start date}

\item{date_to}{end date}

\item{contact_type}{types of contacts e.g. Services,}

\item{min_contact}{minimum number of contacts to be 'valid'}

\item{progress}{object for displaying progress. Leave as \code{NA} if
not required}

\item{progress_detail}{message to report in the 'progress' window}

\item{measure}{some (but not all) \code{report_function} can
choose the 'measure' if not required, then use \code{NA}}

\item{require_type_diabetes}{diabetes methods may give option
to specify diabetes types}

\item{qim_name}{the name of the QIM e.g. 'QIM 01'}

\item{measure_name}{the name of the measure e.g. 'HbA1C'}

\item{state_variable_name}{the name of the column in the return
from \code{report_function} which becomes the 'State' column}

\item{small_cell_suppression}{'suppress' (return zero, or remove
row entirely) if count is less than 5.}

\item{include_all_demographic_groups}{include all possible
demographic groups/subgroups, even if there was no count}

\item{age_min}{minimum age for 'all demographic' option}

\item{age_max}{manimum age for 'all demographic' option}

\item{states}{states for 'all demographic' option}
}
\description{
create reports in a 'standard' format
from the 10 QIM measures

gets report from 'report_function'
then adds extra 'DiabetesType' column if necessary
adds more columns describing the measurement
renames the 'result' column to 'State'

small cell suppression supported
section 2.4 'Small cell suppression' of
"Practice Incentives Program Quality Improvement Measures User Guide"
https://www1.health.gov.au/internet/main/publishing.nsf/Content/
46506AF50A4824B6CA25848600113FFF/$File/PIP\%20QI\%20-\%20User\%20Guide.pdf
}
