% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_65plus.R
\name{list_qim_65plus_appointments}
\alias{list_qim_65plus_appointments}
\title{List of 65 years or older, with Quality Improvement Measures, in the contact list, with appointments}
\usage{
list_qim_65plus_appointments(dMeasure_obj, contact = "contact",
  date_from = NA, date_to = NA, clinicians = NA, min_contact = NA,
  min_date = NA, contact_type = NA, ignoreOld = NA, lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{contact}{types which are accepted. default is "contact"
"contact" chooses the 'contact system dM$list_contact_diabetes ('active' patients)
anything else chooses the 'appointment' system dM$diabetes_list}

\item{date_from}{start date. default is $date_a}

\item{date_to}{end date (inclusive). default is $date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is $contact_min, initially one (1)}

\item{min_date}{most recent contact must be at least min_date. default is $contact_minDate, initially -Inf}

\item{contact_type}{contact types which are accepted. default is $contact_type}

\item{ignoreOld}{ignore results/observatioins that don't qualify for quality improvement measures
if not supplied, reads $qim_ignoreOld}

\item{lazy}{recalculate the 65+ contact list?}
}
\value{
dataframe of Patient (name), InternalID and measures
}
\description{
Filtered by date, and chosen clinicians
}
\details{
QIM 04 - Influenza immunization - most recent. the QIM measure is within last 15 months

the reference date for 'most recent' measurement is 'date_to'
}
