% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_active.R
\name{list_qim_active}
\alias{list_qim_active}
\title{List of active patients, in the contact list}
\usage{
list_qim_active(
  dMeasureQIM_obj,
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  lazy = FALSE,
  store = TRUE
)
}
\arguments{
\item{dMeasureQIM_obj}{dMeasureQIM R6 object}

\item{contact}{patient list. default is $qim_contact.
TRUE chooses the 'contact' system $list_contact_count ('active' patients) from dMeasure object.
FALSE chooses the 'appointment' system $filter_appointments from dMeasure object.}

\item{date_from}{start date. default is $date_a}

\item{date_to}{end date (inclusive). default is $date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is $contact_min, initially one (1)}

\item{min_date}{most recent contact must be at least min_date. default is $contact_minDate, initially -Inf}

\item{max_date}{most recent contact at most max_date. default is $contact_maxDate}

\item{contact_type}{contact types which are accepted. default is $contact_type}

\item{lazy}{recalculate the diabetes contact list?}

\item{store}{keep result in self$qim_active_list}
}
\value{
dataframe of Patient (name), InternalID and demographics
}
\description{
Filtered by date, and chosen clinicians
}
