% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_copd.R
\name{list_qim_copd}
\alias{list_qim_copd}
\title{List of patient with chronic lung disease, with Quality Improvement Measures, in the contact list}
\usage{
list_qim_copd(
  dMeasureQIM_obj,
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE
)
}
\arguments{
\item{dMeasureQIM_obj}{dMeasureQIM R6 object}

\item{contact}{patient list. default is $qim_contact.
TRUE chooses the 'contact' system $list_contact_diabetes ('active' patients) from dMeasure object.
FALSE chooses the 'appointment' system $diabetes_list from dMeasure object.}

\item{date_from}{start date. default is $date_a}

\item{date_to}{end date (inclusive). default is $date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is $contact_min, initially one (1)}

\item{min_date}{most recent contact must be at least min_date. default is $contact_minDate, initially -Inf}

\item{max_date}{most recent contact at most max_date. default is $contact_maxDate}

\item{contact_type}{contact types which are accepted. default is $contact_type}

\item{ignoreOld}{ignore results/observatioins that don't qualify for quality improvement measures
if not supplied, reads $qim_ignoreOld}

\item{lazy}{recalculate the copd contact list?}
}
\value{
dataframe of Patient (name), InternalID, measures
}
\description{
Filtered by date, and chosen clinicians
}
\details{
QIM 06 -Proportion of patients with COPD who were immunised against influenza

the reference date for 'most recent' measurement is 'date_to'
}
