% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_cvdRisk.R
\name{list_qim_cvdRisk}
\alias{list_qim_cvdRisk}
\title{List of patient with information to complete cardiovascular risk assessment}
\usage{
list_qim_cvdRisk(
  dMeasureQIM_obj,
  contact = NA,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = NA,
  lazy = FALSE,
  store = TRUE
)
}
\arguments{
\item{dMeasureQIM_obj}{dMeasureQIM R6 object}

\item{contact}{patient list. default is $qim_contact.

    TRUE chooses the 'contact' system $list_contact_diabetes ('active' patients) from dMeasure object.

    FALSE chooses the 'appointment' system $diabetes_list from dMeasure object.}

\item{date_from}{start date. default is $date_a}

\item{date_to}{end date (inclusive). default is $date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is $contact_min, initially one (1)}

\item{min_date}{most recent contact must be at least min_date. default is $contact_minDate, initially -Inf}

\item{max_date}{most recent contact at most max_date. default is $contact_maxDate}

\item{contact_type}{contact types which are accepted. default is $contact_type}

\item{ignoreOld}{ignore results/observatioins that don't qualify for quality improvement measures
if not supplied, reads $qim_ignoreOld}

\item{lazy}{recalculate the copd contact list?}

\item{store}{keep result in self$qim_cvdRisk_list_appointments}
}
\value{
dataframe of Patient (name), InternalID and measures
}
\description{
Filtered by date, and chosen clinicians
}
\details{
QIM 08.Proportion of patients with the necessary risk factors assessed
 to enable CVD assessment

required parameters

 Age, Ethnicity (especially ATSI status)

  included - Age 45 to 74 years or older

   OR Age 35 or older + ATSI
    (ATSI 35+ optional - included by default - see $qim_cvdRisk_measure)

   Age 75+ excluded
    (Option to include - see $qim_cvdRisk_measure)

 Known cardiovascular disease
    (optional - excluded by default - see $qim_cvdRisk_measure)

 Presence of diabetes. Diabetes and microalbuminuria

 eGFR

 previous diagnosis of familial hypercholesterolaemia

 systolic blood pressure. Diastolic blood pressure

 Serum total cholesterol. Serum HDL cholesterol

source : National Vascular Disease Prevention Alliance (NVDPA) guidelines
https://www.cvdcheck.org.au/australian-absolute-cardiovascular-disease-risk-calculator

the reference date for 'most recent' measurement is 'date_to'
}
