% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_copd.R, R/userInterface.R
\name{qim_copd}
\alias{qim_copd}
\title{dMeasure Quality Improvement Measures - chronic lung disease}
\usage{
qim_copd(input, output, session, dMQIM, contact)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMQIM}{dMeasure QIM R6 object
access to appointments lists, results, correspondence and EMR database}

\item{contact}{(logical) TRUE if using 'contact' list
'contact' list uses active contact methods
FALSE if using 'appointment' list}
}
\value{
none
}
\description{
dMeasure Quality Improvement Measures - chronic lung disease

Quality Improvement COPD - server
}
