% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_cst.R, R/QIM_cst_UI.R
\name{qim_cst}
\alias{qim_cst}
\title{dMeasure Quality Improvement Measures - cervical screening}
\usage{
qim_cst(input, output, session, dMQIM, contact)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMQIM}{dMeasure QIM R6 object
access to appointments lists, results, correspondence and EMR database}

\item{contact}{(logical) TRUE if using 'contact' list
'contact' list uses active contact methods
FALSE if using 'appointment' list}
}
\value{
none
}
\description{
dMeasureQIM - Practice Incentive Program Quality Improvement Measures fields and methods
}
