% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{qim_cvdRisk}
\alias{qim_cvdRisk}
\title{Quality Improvement cardiovascular risk - server}
\usage{
qim_cvdRisk(input, output, session, dMQIM, contact)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMQIM}{dMeasure QIM R6 object
access to appointments lists, results, correspondence and EMR database}

\item{contact}{(logical) TRUE if using 'contact' list
'contact' list uses active contact methods
FALSE if using 'appointment' list}
}
\value{
none
}
\description{
Quality Improvement cardiovascular risk - server
}
