% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_report_UI.R
\name{qim_report_UI}
\alias{qim_report_UI}
\title{dMeasure Quality Improvement Measures - reports for QIM UI

need definitions including 'measure_names'}
\description{
dMeasure Quality Improvement Measures - reports for QIM UI

need definitions including 'measure_names'
}
