% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_reportCharter_UI.R
\name{qim_reportChart}
\alias{qim_reportChart}
\title{dMeasure Quality Improvement Measures - charts for QIM UI

need definitions including 'measure_names'}
\description{
dMeasure Quality Improvement Measures - charts for QIM UI

need definitions including 'measure_names'
}
