% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_active.R
\name{report_qim_active}
\alias{report_qim_active}
\title{Quality Improvement Measure report, in the contact list. Active contacts}
\usage{
report_qim_active(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, min_contact = NA, min_date = NA,
  contact_type = NA, demographic = NA, lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{start date. default is $date_a}

\item{date_to}{end date (inclusive). default is $date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is $contact_min, initially one (1)}

\item{min_date}{most recent contact must be at least min_date. default is $contact_minDate, initially -Inf}

\item{demographic}{demographic groupings for reporting.
if not supplied, reads $qim_demographicGroup
list of available demographic groups in $qim_demographicGroupings}

\item{lazy}{recalculate the diabetes contact list?}
}
\value{
dataframe of Patient (name), demographics, Count, Proportion
}
\description{
Filtered by date, and chosen clinicians
}
