% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{shinydashboardmenuItem}
\alias{shinydashboardmenuItem}
\title{item description for left sidebar menu}
\usage{
shinydashboardmenuItem()
}
\value{
shinydashboard menuItem object
}
\description{
item description for left sidebar menu
}
