% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityImprovementMeasures.R
\name{complete_demographics}
\alias{complete_demographics}
\title{complete demographics}
\usage{
complete_demographics(
  df,
  qim_name,
  age_min = 0,
  age_max = 65,
  include_diabetes = FALSE,
  measure,
  states = c(FALSE, TRUE),
  atsi_only = FALSE
)
}
\arguments{
\item{df}{dataframe to process}

\item{qim_name}{name of qim}

\item{age_min}{minimum age group}

\item{age_max}{maximum age group}

\item{include_diabetes}{include diabetes demographic}

\item{measure}{name of measure}

\item{states}{vector of status types}

\item{atsi_only}{include only ATSI sub-groups?}
}
\value{
dataframe with DOB, Age10, Sex, Indigenous and maybe DiabetesStatus
}
\description{
includes different types of status e.g. TRUE/FALSE
}
