% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{datatableServer}
\alias{datatableServer}
\title{QIM module - server for DailyMeasure/GPstat! web UI}
\usage{
datatableServer(id, dMQIM, contact)
}
\arguments{
\item{id}{id}

\item{dMQIM}{dMeasure QIM R6 object}

\item{contact}{(logical) TRUE if using 'contact' list
'contact' list uses active contact methods
FALSE if using 'appointment' list}
}
\value{
none
}
\description{
QIM module - server for DailyMeasure/GPstat! web UI
}
