% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityImprovementMeasures.R
\docType{data}
\name{dMeasureQIM}
\alias{dMeasureQIM}
\title{dMeasureQIM class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dMeasureQIM
}
\description{
Quality Improvement Measures reports and case-finding in EMR (Best Practice)
    a module for dMeasure
}
\details{
dMeasureQIM class
}
\keyword{datasets}
