% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_reportCharter_UI.R
\name{qim_reportCharter}
\alias{qim_reportCharter}
\title{Quality Improvement report charter - server}
\usage{
qim_reportCharter(input, output, session, dMQIM, report)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMQIM}{dMeasure QIM R6 object
access to appointments lists, results, correspondence and EMR database}

\item{report}{a list returned by qim_reportCreator
should contain $report_values(), which is a dataframe}
}
\value{
none
}
\description{
Quality Improvement report charter - server
}
