% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIM_report.R
\name{writeReportJSON}
\alias{writeReportJSON}
\title{writeReportJSON}
\usage{
writeReportJSON(
  d,
  date_to,
  author_id = "bpsrawdata",
  practice_id,
  product = "GPstat",
  report = "PIP QI",
  version = "1.1",
  small_cell_suppression = TRUE,
  indigenous_aggregate = TRUE,
  sex_aggregate = TRUE
)
}
\arguments{
\item{d}{data}

\item{date_to}{chosen end date}

\item{author_id}{identifier from source software of the authenticated user who
initiated the extract}

\item{practice_id}{an identifier for the practice}

\item{product}{the name of the software that produced the file}

\item{report}{Must be "PIP QI"}

\item{version}{version number of the specification used to generate the JSON}

\item{small_cell_suppression}{'suppress' (return NA, or remove
row entirely) if count is denominator less than 5 OR numerator 2 or less
OR difference between numerator and denomiator is less than or equal to 2.
helps de-identifiability}

\item{indigenous_aggregate}{simplify indigenous groups to
INDIGENOUS, NON-INDIGENOUS, NOT STATED}

\item{sex_aggregate}{simplify sex groups to
MALE, FEMALE, INDETERMINATE/INTERSEX/UNSPECIFIED/NOT STATED/INADEQUATELY DESCRIBED}
}
\value{
JSON string
}
\description{
write reports to PIP QI Eligible Data Set JSON Specification 1.1
}
